fix(security): break codeql cyclic import in context evidence tools
Extract HarnessReport/MatrixRow shared types to context_invocation_harness_types so evidence_json no longer imports the live harness at module level, breaking the CodeQL py/cyclic-import cluster tracked in #3939.
